results

+++ app/pages/D_Investment_reco.py
@@ -104,7 +104,7 @@ for col in stocklist:
     # printing_results.loc[col, 'market_results'] = returns[search_dictio.get(col).upper()].prod() - 1
     st.write(returns[search_dictio.get(col).upper()])
 printing_results['strategy_results'] = printing_results['total_return'] / printing_results['total_investment']
-
+returns[selected_columns] = returns[selected_columns].astype(float)
 st.write(returns[selected_columns].dtypes)
 results = results.rename(columns={'Unnamed: 0': 'companies'})
 results.fillna(0, inplace=True)
